Move webhooks schema to the spec definition
