+++ packages/app/src/cli/models/extensions/specifications/app_config_app_home.ts
@@ -1,4 +1,4 @@
-import {validateUrl} from './configuration/schema.js'
+import {validateUrl} from './validation/common.js'
 import {TransformationConfig, createConfigExtensionSpecification} from '../specification.js'
 import {zod} from '@shopify/cli-kit/node/schema'
